Minor modifications to get CIME to work on Odyssey
Two changes were copied from a working CIME installation to get
CESM2 to run on Odyssey. The first is a minor modification,
correcting the syntax of an error message (utils.py), and may not
be necessary. The second, in config_pio.xml, disables Parallel
NetCDF when running on "odyssey". This is necessary, as otherwise
PIO tries to use Parallel NetCDF and fails to build.

Signed-off-by: Sebastian Eastham <seastham@fas.harvard.edu>